move osu-related (user) to their own folder
